+++ src/main/java/com/controladordefinancas/ControladorDeFinancasApplication.java
@@ -1,4 +1,4 @@
-package com.jedu_lima.controladorDeFinancas;
+package com.controladordefinancas;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
