+++ src/main/java/com/controladordefinancas/controller/DespesaController.java
@@ -1,0 +1,4 @@
+package com.controladordefinancas.controller;
+
+public class DespesaController {
+}
