+++ src/main/java/com/controladordefinancas/entity/enums/Categoria.java
@@ -1,0 +1,4 @@
+package com.controladordefinancas.entity.enums;
+
+public class Categoria {
+}
