+++ src/main/java/com/controladordefinancas/model/CriarDespesaEntradaDTO.java
@@ -1,6 +1,6 @@
-package com.jedu_lima.controladorDeFinancas.model;
+package com.controladordefinancas.model;
 
-import com.jedu_lima.controladorDeFinancas.entity.enums.Categoria;
+import com.controladordefinancas.entity.enums.Categoria;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
