+++ src/main/java/com/controladordefinancas/model/CriarUsuarioEntradaDTO.java
@@ -1,4 +1,4 @@
-package com.jedu_lima.controladorDeFinancas.model;
+package com.controladordefinancas.model;
 
 public record CriarUsuarioEntradaDTO(Long id, String login, String senha) {
 }
