+++ src/main/java/com/controladordefinancas/repository/UsuarioRepository.java
@@ -1,6 +1,6 @@
-package com.jedu_lima.controladorDeFinancas.repository;
+package com.controladordefinancas.repository;
 
-import com.jedu_lima.controladorDeFinancas.entity.UsuarioEntity;
+import com.controladordefinancas.entity.UsuarioEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface UsuarioRepository extends MongoRepository<UsuarioEntity, Long> {
