+++ src/main/java/com/controladordefinancas/service/DespesaService.java
@@ -1,0 +1,4 @@
+package com.controladordefinancas.service;
+
+public interface DespesaService {
+}
