+++ src/main/java/com/controladordefinancas/service/impl/DespesaServiceImpl.java
@@ -1,0 +1,4 @@
+package com.controladordefinancas.service.impl;
+
+public class DespesaServiceImpl {
+}
