--- src/main/java/com/jedu_lima/controladorDeFinancas/controller/DespesaController.java
@@ -1,4 +1,0 @@
-package com.jedu_lima.controladorDeFinancas.controller;
-
-public class DespesaController {
-}
