--- src/main/java/com/jedu_lima/controladorDeFinancas/controller/UsuarioController.java
@@ -1,4 +1,0 @@
-package com.jedu_lima.controladorDeFinancas.controller;
-
-public class UsuarioController {
-}
