--- src/main/java/com/jedu_lima/controladorDeFinancas/entity/enums/Categoria.java
@@ -1,4 +1,0 @@
-package com.jedu_lima.controladorDeFinancas.entity.enums;
-
-public class Categoria {
-}
