--- src/main/java/com/jedu_lima/controladorDeFinancas/service/DespesaService.java
@@ -1,4 +1,0 @@
-package com.jedu_lima.controladorDeFinancas.service;
-
-public interface DespesaService {
-}
