--- src/main/java/com/jedu_lima/controladorDeFinancas/service/impl/UsuarioServiceImpl.java
@@ -1,4 +1,0 @@
-package com.jedu_lima.controladorDeFinancas.service.impl;
-
-public class UsuarioServiceImpl {
-}
